+++ docs/CHANGELOG.md
@@ -30,6 +30,8 @@ Here is a list of changes for each of the releases.
 - Rules
 
   - Added workaround rules for `Microsoft Teams` app present in Windows 11
+  - Added rule for Network List Service
+  - Updated DNS rule to include port for encrypted DNS on Windows 11
 
 - Modules